feat: raw mode implementation
