+++ packages/@tinacms/icons/src/index.ts
@@ -50,6 +50,7 @@ export * from './Lock'
 export * from './Quote'
 export * from './Underline'
 export * from './Strikethrough'
+export * from './Markdown'
 export * from './Alert'
 export * from './Info'
 export * from './Warning'
