+++ packages/react-tinacms-editor/src/plugins/Inline/index.ts
@@ -17,4 +17,4 @@ limitations under the License.
 */
 
 export { inlinePlugin } from './plugin'
-export { Menu } from './Menu'
+export * from './Menu'
